+++ src/inbuilt-commands/stats.ts
@@ -1,16 +1,16 @@
 import Command, { PermissionLevel } from "../commands/command"
 import { Client } from "../core/client"
 import DisharmonyMessage from "../models/discord/disharmony-message"
+import { RichEmbed } from "discord.js"
 
-function invoke(_: string[], __: DisharmonyMessage, client: Client)
-{
-    return Promise.resolve(
-        `
-        **Server count:** ${client.stats.guildCount}
-        **Cached users:** ${client.stats.userCount}
-        **Uptime:** ${client.stats.uptimeStr}
-        `,
-    )
+async function invoke(_: string[], message: DisharmonyMessage, client: Client): Promise<void> {
+    const embed = new RichEmbed()
+        .setTitle(`__${client.config.serviceName} stats__`)
+        .addField("Server Count", client.stats.guildCount, false)
+        .addField("Cached users", client.stats.userCount, false)
+        .addField("Uptime", client.stats.uptimeStr, false)
+
+    await message.reply(embed)
 }
 
 export default new Command(